feat: upgrade to node 22 (#194)
BREAKING CHANGE: upgrade to node 22

Additionally added support for zod output generation, including a command to migrate existing therefore schemas to zod usint the --migrate-to=zod option on the cli. This will also generate an interop to make sure most code will work as is, with required intention to fully move towards zod as schema language.